Update gitcommitSummary color
This will update the `gitcommitSummary` color to blue.

The `gitcommitSummary` is defined as a Keyword:
https://github.com/vim/vim/blob/master/runtime/syntax/gitcommit.vim#L64

In the One Half theme, Keywords are set to be red. The problem with
red is when spellchecking is enabled, misspelled words are also set
to be red. (Yellow is also not a good choice because there are other
spelling-related issues that are set to be yellow.)