vim: resolve misleading bg color of split buffers
When two or more buffers are open at the same time, the one without
the focus has a blue background which, being more visible, can make
the user think that this is the on-focus buffer, while it is not.

Attempt fix setting Pmenu colors to have the same bg color of the on
focus split buffer, while using a grayed fg color.