Introduce color for whitespace markers in Neovim.
This is a Neovim only feature, see
https://github.com/neovim/neovim/pull/6367

When using

set list

In Vim, the whitespace characters would be drawn in the same color as
the text which is intrusive IMHO. This commit changes it to use the same
color as comments.

In order to make this compatible with regular Vim, one would probably
have to use `NonText`, instead.