Update README for modern vim ecosystem; fix lightline example (#87)
Listing existing plugins is superfluous, as their users know them. Some of the listed ones are no longer maintained. vim-plug (and maybe dein.vim) appear to be the most popular ones, and the former is what most projects use in examples.

The lightline example is slightly changed, as the `g:lightline` variable cannot be assumed to exist.